+++ SpectatorFootball/Game/Play_Struct.cs
@@ -34,7 +34,7 @@ namespace SpectatorFootball.GameNS
         public string Long_Message;
 
         public Double Line_of_Scimmage;
-        public int Vertical_Ball_Placement;
+        public Double Vertical_Ball_Placement;
 
         public Play_Package Offensive_Package = null;
         public Formation Defensive_Formation;
